+++ docs/pages/components/modal/Modal.vue
@@ -55,7 +55,7 @@
                 ExFullScreenCode,
                 programmaticSyntax: `
                 // From inside Vue instance
-                this.$modal.open(props)
+                this.$buefy.modal.open(props)
 
                 // From outside Vue instance
                 import { ModalProgrammatic as Modal } from 'buefy'