+++ docs/pages/extensions/veevalidate/examples/ExSimple.vue
@@ -105,14 +105,14 @@
             validateBeforeSubmit() {
                 this.$validator.validateAll().then((result) => {
                     if (result) {
-                        this.$toast.open({
+                        this.$buefy.toast.open({
                             message: 'Form is valid!',
                             type: 'is-success',
                             position: 'is-bottom'
                         })
                         return;
                     }
-                    this.$toast.open({
+                    this.$buefy.toast.open({
                         message: 'Form is not valid! Please check the fields.',
                         type: 'is-danger',
                         position: 'is-bottom'