+++ src/components/modal/index.js
@@ -32,7 +32,7 @@ const ModalProgrammatic = {
 const Plugin = {
     install(Vue) {
         registerComponent(Vue, Modal)
-        registerComponentProgrammatic(Vue, '$modal', ModalProgrammatic)
+        registerComponentProgrammatic(Vue, 'modal', ModalProgrammatic)
     }
 }